components配下に修正した。
